requirements.txt
prettytable
requests
python-dotenv
IPython